DebatesDjango/requirements.txt
django
django-rest-framework
django-guardian
